Remove BIGSI and SBT code
